AMQ: split platform-specific wake helpers
